buf_copy: remove thin wrapper & useless conversion
